src/python-azure-ai-analyze-text/main.py
from dotenv import load_dotenv
import os
from pathlib import Path

# Import namespaces
from azure.core.credentials import AzureKeyCredential
from azure.ai.textanalytics import TextAnalyticsClient


def main():

    try:
        # Get Configuration Settings
        ENV_PATH = Path(__file__).resolve().parent / ".env"  # Path to .env file
        load_dotenv(dotenv_path=ENV_PATH)  # Load .env file

        cog_endpoint = os.getenv("COG_SERVICE_ENDPOINT")
        cog_key = os.getenv("COG_SERVICE_KEY")

        # Create client using endpoint and key
        credential = AzureKeyCredential(cog_key)
        cog_client = TextAnalyticsClient(endpoint=cog_endpoint, credential=credential)

        # Analyze each text file in the reviews folder
        reviews_folder = "reviews"
        BASE_DIR = Path(__file__).resolve().parent  # Get the current directory
        reviews_folder = (
            BASE_DIR / reviews_folder
        )  # Create the full path to the reviews folder

        print("CWD:", os.getcwd())  # Print the current working directory for debugging
        print(
            "Reviews Folder:", reviews_folder
        )  # Print the reviews folder path for debugging

        for file_path in reviews_folder.iterdir():
            # Read the file contents
            if file_path.is_file():
                print("\n-------------\n" + file_path.name)
                text = file_path.read_text(encoding="utf8")
                print("\n" + text)

                # Get language
                detectedLanguage = cog_client.detect_language(documents=[text])[0]
                print("\nLanguage: {}".format(detectedLanguage.primary_language.name))

                # Get sentiment
                sentimentAnalysis = cog_client.analyze_sentiment(documents=[text])[0]
                print("\nSentiment: {}".format(sentimentAnalysis.sentiment))

                # Get key phrases
                phrases = cog_client.extract_key_phrases(documents=[text])[
                    0
                ].key_phrases
                if len(phrases) > 0:
                    print("\nKey Phrases:")
                    for phrase in phrases:
                        print("\t{}".format(phrase))

                # Get entities
                entities = cog_client.recognize_entities(documents=[text])[0].entities
                if len(entities) > 0:
                    print("\nEntities")
                    for entity in entities:
                        print("\t{} ({})".format(entity.text, entity.category))

                # Get linked entities
                entities = cog_client.recognize_linked_entities(documents=[text])[
                    0
                ].entities
                if len(entities) > 0:
                    print("\nLinks")
                    for linked_entity in entities:
                        print("\t{} ({})".format(linked_entity.name, linked_entity.url))

    except Exception as ex:
        print(ex)


if __name__ == "__main__":
    main()
