pyproject.toml
[project]
name = "python-azure-ai-analyze-text"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "azure-ai-textanalytics (>=5.3,<6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
flake8 = "^7.3.0"
pre-commit = "^4.5.1"
python-dotenv = "^1.2.1"


[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"


[tool.poetry]
packages = [{ include = "python-azure-ai-analyze-text", from = "src" }]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.flake8]
max-line-length = 88
ignore = ["E203", "W503"]
exclude = [".venv", "__pycache__", "build", "dist"]
